Introduce AiO HotPlug v1.0
Signed-off-by: Shoaib0597 <Shoaib0595@gmail.com>

An All in One (AiO) HotPlug for Octa-Core big.LITTLE SoCs. The HotPlug is Based on the Work (State Helper Driver) of neobuddy89.

This HotPlug aims to provide All of the Functionality which a user wants from a HotPlug, ranging from Normal Offline/Online of Cores to Completely Offlining Entire Clusters. As the name suggests, this HotPlug is a Complete Package.

Signed-off-by: Panchajanya1999 <rsk52959@gmail.com>